feat: windows gleam install and run support
To support this a few things change with where installations are
done. A better following of XDG patterns in general is included
as well as proper placement on mac and windows.

Now on Linux the hard links like `gleam` and `erl` are in the dir
`~/.local/bin`. On Mac and Windows these are found in the dir
`~/.beamup/bin`. The configuration file `config.toml` stays the
same on linux, `~/.config/beamup/config.toml` but on Mac it is now
`~/Library/Application Support/beamup/config.toml` and on Windows
`~\AppData\Local\beamup\config.toml`. The installations of langs
and components on Linux is done to `~/.local/share/beamup`, on
Mac to `~/Library/Application Support/beamup` and on Windows to
`~\AppData\Local\beamup`.