fix: don't crash when running releases command on elixir
this would happen because releases command creates a sort of mock Elixir
struct with an empty id that beamup then tried to find the major
otp version for which would fail. A better fix is needed to not
be setting up these fake values to make things succceed when
they don't need to.